requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
future==0.18.2
gast==0.3.3
geomloss==0.2.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
importlib-metadata==2.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
multidict==4.7.6
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
POT==0.7.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pykeops==1.4
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
termcolor==1.1.0
tifffile==2020.9.3
torch==1.4.0
torchvision==0.5.0
typeguard==2.9.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
zipp==3.3.0
